tests/wpt/tests/html/dom/render-blocking/element-render-blocking-017.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Adding blocking attr in the body has no effect</title>

<link id=link rel=expect href="#last">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "adding 'blocking=render' in the body has no effect");
</script>
</head>
<body>
<script>
link.blocking = "render"
</script>
  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
